refactor: 불필요한 코드 삭제 및 폴더명 일관성 있게 변경
