+++ src/features/recruitments/RecruitmentInfo/index.tsx
